+++ src/biodwh2-hpo/src/main/java/de/unibi/agbi/biodwh2/hpo/etl/HPOGraphExporter.java
@@ -1,12 +1,9 @@
 package de.unibi.agbi.biodwh2.hpo.etl;
 
-import de.unibi.agbi.biodwh2.core.Workspace;
-import de.unibi.agbi.biodwh2.core.etl.GraphExporter;
-import de.unibi.agbi.biodwh2.core.exceptions.ExporterException;
-import de.unibi.agbi.biodwh2.core.model.graph.Graph;
+import de.unibi.agbi.biodwh2.core.etl.OntologyGraphExporter;
 import de.unibi.agbi.biodwh2.hpo.HPODataSource;
 
-public class HPOGraphExporter extends GraphExporter<HPODataSource> {
+public class HPOGraphExporter extends OntologyGraphExporter<HPODataSource> {
     public HPOGraphExporter(final HPODataSource dataSource) {
         super(dataSource);
     }
@@ -17,7 +14,7 @@ public class HPOGraphExporter extends GraphExporter<HPODataSource> {
     }
 
     @Override
-    protected boolean exportGraph(final Workspace workspace, final Graph graph) {
-        return true;
+    protected String getOntologyFileName() {
+        return "hp.obo";
     }
 }